Suppression de Javascript et Fichier non utilisé
