--- backend/src/config/config.exemple.js
@@ -1,19 +1,0 @@
-/**
- * @source backend/src/config
- * Ce gabarit de configuration centralisee sert uniquement de reference documentaire.
- * Indiquer dans ce module les valeurs modifiable via variables d'environnement
- * ainsi que les dependances critiques pour le backend Vino.
- *
- * Guide rapide:
- * 1. Dupliquer ce fichier sous un nom explicite.
- * 2. Renseigner les valeurs concretes dans chaque cle.
- * 3. Documenter les decisions critiques pour garder une trace historique.
- */
-export const configurationExemple = {
-  // nomService: 'vino-backend',
-  // portHttp: 3000,
-  // urlBaseDonnees: 'postgresql://utilisateur:motdepasse@hote:5432/vino',
-  // secrets: {
-  //   cleApiSaq: 'remplacer-par-variable-env',
-  // },
-};
