+++ frontend/src/pages/Catalogue.jsx
@@ -5,7 +5,7 @@ import MenuEnBas from "@components/components-partages/MenuEnBas/MenuEnBas";
 import CarteBouteille from "@components/carte/CarteBouteille";
 import Message from "@components/components-partages/Message/Message";
 import NonTrouver from "@components/components-partages/NonTrouver/NonTrouverLazy";
-import Spinner from "@components/components-partages/Spinner/SpinnerSimple";
+import Spinner from "@components/components-partages/Spinner/Spinner";
 import Filtres from "@components/components-partages/Filtre/Filtre";
 
 import authentificationStore from "@store/authentificationStore";
